Wait until client is connected
If sending after client exists but before it is connected (or while it
is reconnecting) then calling client.send will throw an error and not
save anything into SM state. So push these stanzas into pending to be
sent when coming online.

Not that this is not part of SM state and is not persisted at all, so a
full crash/reset will lose these stanzas right now, unlike un-acked SM stanzas.