Queue outgoing audio packets
With RTP, sending too fast or too slow can cause playback issues for the
receiver. So we queue up a bit (start at a 100ms buffer and growing to a
max of 5000ms depending on if we ever run out) and send it out at a
constant rate with a timer, to smooth over any timing bumps in the
production of any single packet.

If packets come too fast we stop asking for them until the queue shrinks
again.

If some packets come too slow we keep sending out of the queue so long
as we have any. If every single packet takes more than its length to
produce you'll be forever behind and get weird lag issues still.