New media strategy
Return ni:/// URIs (or, optionally, /.well-known/ni/ paths) for media.
Provide a helper to serve the .well-known paths from a service worker.

No more option to get media URIs from the persistence layer, instead
each layer will provide a fit-for-platform mechanism to map ni URIs to
data, file paths, whatever make sense in context. For web that's
Response object or service worker interception.

Avatars are now two-part: the possible-null ni URI and the never-null
placeholder. Display the placeholder while loading the real avatar data,
or if it is null or loading it fails for any reason.