Store call history in with message history
This design allows for a kind of message in chats that represents a
call.  ChatMessage#type == MessageCall

This type of message can be edited by either side, and the top level
localId is set to the sessionid.  This means that the whole history of
call status changes ends up in ChatMessage#versions

A helper was added to Client to allow easily storing a message that
way (without direct access to persistence which you need to know the
accountId for anyway), we should probably use that more places now.

Some events are stored even though we don't send them. Namely ringing
and retract. We may decide to send them in some cases eventually, and
they're useful for the local data model so we store as if we sent them already.