Support for blocking/unblocking chats
Block status is persisted on the chat.
It is updated when server pushes, and a full blocklist pull is done on
new session.
Block closes any open chat and prevents any notifications for that chat
so it should not notify or open again even on MAM.
Explicit unblock opens the chat.

Also implemented spam reporting when making the block call if the app
wishes to do so. SPAM reports must contain an example SPAM message.