OMEMO: Refactor encryption status reporting
Previously we put the encryption status into the stanza, as this was a
convenient way to pass the information around with the stanza. A downside is
that this could be forged by a sender (or intermediate hop) unless we make
sure to always strip it. It was also a bit more clunky and less performant
than simply creating and passing around an in-memory structure. So that's what
we do now - though some API function signatures have necessarily changed.