Fix first start race
The problem here is that we don't wait for this to be done before moving
on to other parts of the sync. So this may complete after the general
chats/update for all known chats has happened and even after the
joinAllChannels has happened. So make sure we update the UI and selfPing
in case we aren't joined.