Force ordered delivery of stanzas while live
There can be async prep work to do before sending a stanza. Force
stanzas to go in order even if their prep work takes different amounts
of time.

The outbox can also be paused so that stanzas can be delayed eg for when
a MUC is still joining and it is not safe to send a message yet.