Invited uistate
For presence requests and MUC invites, set to "Invited" which means not
quite Open but not quite Closed. Appears in getChats just like
Open/Pinned but MUC doesn't join, etc.

Invited transitions to Open when you send or receive a message or call,
or set trusted or bookmark the chat. Transitions to closed when you call close.